Ragdoll info for role
I added "OldRoleId" in player to have the last role Id of the player. This is on all useful when a player is dead.

I added "RoleType" in the Ragdoll for get the RoleType of the Ragdoll.

I added "RoleId" is the roll id of the player. it use "OldRoleId".
it might be useful for a future update. This would prevent SCP 049 from resuscitating humanoid SCPs such as SCP 035 and SCP 056 as they are from the SCP team.